+++ src/main.c
@@ -1,10 +1,10 @@
 //#define DEVMODE
-//#define START
 
 #include <stdio.h>
 #include <SDL2/SDL.h>
 #include "devm.h"
 #include "disp.h"
+#include "file.h"
 #include "harv.h"
 #include "hold.h"
 #include "loot.h"
@@ -19,8 +19,6 @@
 /* Function prototypes */
 static void	game_init(void);
 static void	game_quit(void);
-static void	generate_farts(struct game *cur_game, struct worldmap *main_map);
-static void	copy_fart(struct worldmap *main_map, struct worldmap *fart, int row, int col);
 
 /* Global game construct */
 struct game GAME = {
@@ -50,6 +48,10 @@ main()
 	/* initialize game */
 	game_init();
 	
+	/* go to the title screen */
+	title_screen(&GAME, &MAP, &PLAYER, STARTING_GAME);
+	if (GAME.running == SDL_FALSE) return 0;
+	
 	/* draw map, player, render and "move player" to update map */
 	draw_all(&GAME, &MAP, &PLAYER);
 	move_player(&GAME, &MAP, &PLAYER, 0, 0);
@@ -73,7 +75,7 @@ main()
 		} else if (event.type == SDL_KEYDOWN) {
 			switch (event.key.keysym.sym) {
 				case SDLK_ESCAPE: /* quit */
-					GAME.running = SDL_FALSE;
+					title_screen(&GAME, &MAP, &PLAYER, GAME_IN_PROGRESS);
 					break;
 				case SDLK_UP: /* move up */
 				case SDLK_w:
@@ -153,7 +155,6 @@ main()
 		}
 	}
 	/* quit game and exit normally */
-	save_all(&MAP, &PLAYER);
 	game_quit();
 	return 0;
 }
@@ -165,21 +166,6 @@ game_init(void)
 	display_init(&GAME);
 	/* Seed RNG */
 	seed_rng();
-	#ifdef START
-		/* Set up the worldmap */
-		create_map(&MAP, MAP_ROWS, MAP_COLS);
-		/* Generate farts and copy to worldmap */
-		generate_farts(&GAME, &MAP);
-		/* Set up player */
-		player_init(&MAP, &PLAYER);
-		/* Set up depleted item table */
-		setup_dtable();
-	#else
-		load_all(&GAME, &MAP, &PLAYER);
-	#endif
-	/* Set up holder table */
-	setup_hold();
-	/* The window is now up and running */
 	GAME.running = SDL_TRUE;
 }
 
@@ -196,63 +182,4 @@ game_quit(void)
 	free_map(&MAP);
 	/* Quit SDL display functions */
 	display_quit(&GAME);	
-}
-
-static void
-generate_farts(struct game *cur_game, struct worldmap *main_map)
-{
-	int rows, cols;
-	struct worldmap biomes;
-	struct worldmap fart;
-
-	/* Pick random biomes */
-	create_map(&biomes, 32, 64);
-	populate_map(&biomes, 1, 5);
-	
-	/* Create farts */
-	for (rows = 0; rows < 32; rows++) {
-		/* Loading bar */
-		loading_bar(cur_game, "Generating maps", 100*rows/32);
-		for (cols = 0; cols < 64; cols++) {
-			/* Create a fart */
-			create_map(&fart, 24, 24);
-			/* Populate with a random biome */
-			populate_map(&fart, 1, *(*(biomes.tile+rows)+cols) - 1);
-			/* copy_fart */
-			copy_fart(main_map, &fart, rows * 16 - 4, cols * 16 - 4);
-			/* free_map */
-			free_map(&fart);
-		}
-	}
-	/* Free the biome plan */
-	free_map(&biomes);
-}
-
-static void
-copy_fart(struct worldmap *main_map, struct worldmap *fart, int row, int col)
-{
-	int rows, cols;
-	
-	/* Make sure it fits in map */
-	if (row < 0) {
-		fart->row_size += row;
-		row = 0;
-	}
-	if (col < 0) {
-		fart->col_size += col;
-		col = 0;
-	}
-	
-	for (rows = 0; rows < fart->row_size; rows++) {
-		for (cols = 0; cols < fart->col_size; cols++) {
-			/* Make sure you're within the map */
-			if (rows + row > main_map->row_size - 1) continue;
-			if (cols + col > main_map->col_size - 1) continue;
-			/* If there's an existing tile, blend it maybe */
-			if (*(*(main_map->tile+rows+row)+cols+col) == 0 || rand_num(0, 1) == 1) {
-				*(*(main_map->tile+rows+row)+cols+col) = *(*(fart->tile+rows)+cols);
-				*(*(main_map->biome+rows+row)+cols+col) = *(*(fart->biome+rows)+cols);
-			}
-		}
-	}
 }